refactor(notes): extract sidebar JSX into NotesSidebar component
Extract the 920-line sidebar <aside> block from NotesManagerPage into
a dedicated React.memo-wrapped NotesSidebar component with typed props.
All 137 tests pass unchanged. NotesManagerPage: 7852 → 7022 lines.
